Full Results and Analysis for Results
MC now has loop over number of lands in deck and saves stats for each of them. Also created analysis script that now makes a table of the MC results. Need to add plot with sums of 2-4 lands in opener etc. Then make histos. Then maybe compute odds of 3 lands on 3, 4 on 4, etc.