Finished Analytic Calculations
Finished analytically calculating the number of lands in your opening hand odds for normal and best of one hands. Added analysis file to put the results in a table. Now working on the odds of hitting x lands on turn x. Works for 1 land in the opening hand but not for 2 or 3 yet.